+++ changelog.rst
@@ -1,7 +1,13 @@
 Changelog
 =========
 
-Last change: 16-JAN-2026 CEST
+Last change: 12-JAN-2026 CEST
+
+G5M
+---
+- Added G5M
+- Axial localization precision
+- Subclustering test in picasso.clusterer and Filter and after g5m in render
 
 0.9.4
 -----